2024/08/02

+++ zhuban_new/Core/Inc/adc.h
@@ -49,12 +49,14 @@ void MX_ADC1_Init(void);
 void MX_ADC3_Init(void);
 
 /* USER CODE BEGIN Prototypes */
+
+
 typedef struct {
     uint32_t *buffer;
     uint32_t channelCount;
 } ADCMessage;
 
-void CalculateAverage(uint32_t *buffer, uint32_t *average, uint32_t channelCount, uint32_t sampleCount);
+void CalculateAverage(uint32_t *buffer, float *average, uint32_t channelCount, uint32_t sampleCount);
 
 void Adc_conv_start(void);
